+++ test/SpagoSpec.hs
@@ -692,6 +692,16 @@ spec = around_ setup $ do
       spago ["build"] >>= shouldBeSuccess
       spago ["run", "--exec-args", "hello world", "--node-args", "hallo welt"] >>= shouldBeSuccessOutput "run-args-combined-output.txt"
 
+  describe "spago script" $ do
+
+    -- TODO: Once 0.14 of the compiler is released, compilation logs will be sent to `stderr`
+    -- At that time, we should add more tests checking the output of the script
+
+    it "Spago script should create file in directory where it is executed" $ do
+
+      spago ["script", "../fixtures/spago-script-make-file.purs", "-d", "node-fs"] >>= shouldBeSuccess
+      checkFixture "spago-script-result.txt"
+
   describe "spago bundle" $ do
 
     it "Spago should fail but should point to the replacement command" $ do